+++ packages/frontend/rewrite/src/pages/about/getAboutUsData.ts
@@ -1,6 +1,6 @@
 import { getAppLayoutProps } from 'rewrite/src/common/getAppLayoutProps'
 import { getMetadata } from 'rewrite/src/ssr/head/getMetadata'
-import type { RenderData } from 'rewrite/src/ssr/server'
+import type { RenderData } from 'rewrite/src/ssr/types'
 import type { Manifest } from '~/utils/Manifest'
 
 export async function getAboutUsData(
