+++ packages/frontend/rewrite/src/pages/da-risk-framework/DaRiskFrameworkRouter.ts
@@ -1,5 +1,5 @@
 import type { Router } from 'express'
-import type { RenderFunction } from 'rewrite/src/ssr/server'
+import type { RenderFunction } from 'rewrite/src/ssr/types'
 import type { Manifest } from '~/utils/Manifest'
 import { getDaRiskFrameworkData } from './getDaRiskFrameworkData'
