+++ packages/frontend/rewrite/src/pages/da-risk-framework/getDaRiskFrameworkData.ts
@@ -1,5 +1,5 @@
 import { getMetadata } from 'rewrite/src/ssr/head/getMetadata'
-import type { RenderData } from 'rewrite/src/ssr/server'
+import type { RenderData } from 'rewrite/src/ssr/types'
 import type { Manifest } from '~/utils/Manifest'
 
 export function getDaRiskFrameworkData(
