+++ packages/frontend/rewrite/src/pages/data-availability/summary/getDataAvailabilitySummaryData.ts
@@ -1,6 +1,6 @@
 import { getAppLayoutProps } from 'rewrite/src/common/getAppLayoutProps'
 import { getMetadata } from 'rewrite/src/ssr/head/getMetadata'
-import type { RenderData } from 'rewrite/src/ssr/server'
+import type { RenderData } from 'rewrite/src/ssr/types'
 import { getDaSummaryEntries } from '~/server/features/data-availability/summary/get-da-summary-entries'
 import { getDaThroughputSummary } from '~/server/features/data-availability/throughput/get-da-throughput-summary'
 import type { Manifest } from '~/utils/Manifest'
