+++ packages/frontend/rewrite/src/pages/ecosystems/EcosystemsRouter.tsx
@@ -1,9 +1,9 @@
 import type { Router } from 'express'
+import type { RenderFunction } from 'rewrite/src/ssr/types'
 import { validateRoute } from 'rewrite/src/utils/validateRoute'
 import { z } from 'zod'
 import { env } from '~/env'
 import type { Manifest } from '../../../../src/utils/Manifest'
-import type { RenderFunction } from '../../ssr/server'
 import { getEcosystemProjectData } from './project/getEcosystemProjectData'
 
 export function EcosystemsRouter(
