+++ packages/frontend/rewrite/src/pages/governance/GovernanceRouter.ts
@@ -1,8 +1,8 @@
 import type { Router } from 'express'
+import type { RenderFunction } from 'rewrite/src/ssr/types'
 import { validateRoute } from 'rewrite/src/utils/validateRoute'
 import { z } from 'zod'
 import type { Manifest } from '../../../../src/utils/Manifest'
-import type { RenderFunction } from '../../ssr/server'
 import { getGovernanceData } from './getGovernanceData'
 import { getGovernancePublicationData } from './publication/getGovernancePublicationData'
 import { getGovernancePublicationsData } from './publications/getGovernancePublicationsData'
