+++ packages/frontend/rewrite/src/pages/governance/publications/getGovernancePublicationsData.tsx
@@ -2,7 +2,7 @@ import { getCollection } from '~/content/get-collection'
 
 import { getAppLayoutProps } from 'rewrite/src/common/getAppLayoutProps'
 import { getMetadata } from 'rewrite/src/ssr/head/getMetadata'
-import type { RenderData } from 'rewrite/src/ssr/server'
+import type { RenderData } from 'rewrite/src/ssr/types'
 import { getGovernancePublicationEntry } from '~/app/(side-nav)/governance/_utils/get-governance-publication-entry'
 import type { Manifest } from '~/utils/Manifest'
