+++ packages/frontend/rewrite/src/pages/scaling/activity/getScalingActivityData.ts
@@ -1,7 +1,7 @@
 import { HOMEPAGE_MILESTONES } from '@l2beat/config'
 import { getAppLayoutProps } from 'rewrite/src/common/getAppLayoutProps'
 import { getMetadata } from 'rewrite/src/ssr/head/getMetadata'
-import type { RenderData } from 'rewrite/src/ssr/server'
+import type { RenderData } from 'rewrite/src/ssr/types'
 import { getScalingActivityEntries } from '~/server/features/scaling/activity/get-scaling-activity-entries'
 import type { Manifest } from '~/utils/Manifest'
