+++ packages/frontend/rewrite/src/pages/zk-catalog/ZkCatalogRouter.ts
@@ -1,8 +1,8 @@
 import type { Router } from 'express'
+import type { RenderFunction } from 'rewrite/src/ssr/types'
 import { validateRoute } from 'rewrite/src/utils/validateRoute'
 import { z } from 'zod'
 import type { Manifest } from '../../../../src/utils/Manifest'
-import type { RenderFunction } from '../../ssr/server'
 import { getZkCatalogData } from './getZkCatalogData'
 import { getZkCatalogProjectData } from './project/getZkCatalogProjectData'
