+++ packages/frontend/rewrite/src/ssr/client-entry.tsx
@@ -9,6 +9,7 @@ declare global {
 }
 
 // biome-ignore lint/style/noNonNullAssertion: It's there
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 const root = document.getElementById('root')!
 hydrateRoot(
   root,
